refactor(ai): 拆出 AIProviderQuickEditOverlay 表单逻辑 hook 与各子面板
